requirements.txt
flask
numpy
opencv-python
face-recognition
pygame
pyserial
